refactor: extract TimerState and helpers from timer_provider
Move TimerPhase enum and TimerState class into a dedicated
timer_state.dart; re-export from timer_provider.dart so all
existing consumer imports remain unchanged.
